backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yt-dlp>=2024.10.0
groq>=0.11.0
ffmpeg-python>=0.2.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
cloudinary>=1.40.0
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.12
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
